requirements.txt
# LLM
google-genai>=1.0.0

# Vector store
qdrant-client>=1.9.0

# Knowledge graph
neo4j>=5.20.0

# Orchestration
langgraph>=0.2.0
langchain-google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0